Only optimize scale on configured repos
It worked fine and improved things in my testing with a repo configured
to use the new settings. On default/old repos however,it can cause minor
issues somehow. Since it's only a small optimization, it might be better
to just remove the whole idea again, or find a better solution.
I'll leave that as a decision for later.