Remove adjusted scaling
Reverts back to original scale calcs, which end up a bit smaller
but tend to display less overrun issues due to this.

Cases where max lines of text can be configured down to 1
and scaling is needed will need to switch functions depending
on this value, as shown in this commit for skill names.